@@ -56,44 +56,45 @@
     "@types/cookie": "^0.3.2",
     "@types/hapi": "^17.8.2",
     "@types/hapi-auth-jwt2": "^8.0.2",
-    "@types/jest": "^23.3.12",
+    "@types/jest": "^24.0.9",
     "@types/jsonwebtoken": "^8.3.0",
     "@types/minimist": "^1.2.0",
-    "@types/pdfmake": "^0.1.2",
-    "@types/ramda": "^0.25.46",
+    "@types/pdfmake": "^0.1.3",
+    "@types/ramda": "^0.25.51",
     "axios": "^0.18.0",
-    "bcrypt": "^3.0.3",
+    "bcrypt": "^3.0.4",
     "boom": "^7.3.0",
     "chalk": "^2.4.2",
     "cookie": "^0.3.1",
+    "csv": "^5.1.1",
     "dotenv": "^6.2.0",
     "hapi": "^17.8.1",
     "hapi-auth-bearer-token": "^6.1.1",
-    "hapi-auth-jwt2": "^8.2.0",
-    "hapi-pino": "^5.2.0",
+    "hapi-auth-jwt2": "^8.3.0",
+    "hapi-pino": "^5.4.1",
     "hapi-qs": "^2.0.1",
     "joi": "^14.3.1",
-    "jsonwebtoken": "^8.4.0",
+    "jsonwebtoken": "^8.5.0",
     "knex": "^0.16.3",
     "minimist": "^1.2.0",
-    "moment": "^2.23.0",
-    "pdfmake": "^0.1.39",
+    "moment": "^2.24.0",
+    "pdfmake": "^0.1.53",
     "pg": "^7.7.1",
     "postmark": "^2.2.4",
-    "qrcode": "^1.3.2",
+    "qrcode": "^1.3.3",
     "ramda": "^0.26.1",
-    "ts-node": "^7.0.1",
-    "typescript": "^3.2.2"
+    "ts-node": "^8.0.2",
+    "typescript": "^3.3.3333"
   },
   "devDependencies": {
     "@types/mockdate": "^2.0.0",
     "axios-mock-adapter": "^1.16.0",
     "factory-girl": "^5.0.4",
-    "jest": "^23.6.0",
+    "jest": "^24.1.0",
     "minimist": "^1.2.0",
     "mockdate": "^2.0.2",
-    "nodemon": "^1.18.9",
-    "ts-jest": "^23.10.5",
+    "nodemon": "^1.18.10",
+    "ts-jest": "^24.0.0",
     "tslint-config-twine": "github:TwinePlatform/tslint-config-twine#2.1.0"
   }
 }
