@@ -30,8 +30,7 @@ exports.migrate = {
 
     const date = (new Date()).toISOString().slice(0, 10).replace(/-/g, '');
     const files = fs.readdirSync(path.resolve(__dirname, 'migrations', 'sql')).sort();
-    const newVersion = Number(last(files).split('_')[1]) + 1;
-    const newFilename = `${date}_${newVersion}_${name.toLowerCase().replace(/\s/g, '_')}`;
+    const newFilename = `${date}_${name.toLowerCase().replace(/\s/g, '_')}`;
     write(path.resolve(MIGRATIONS_BASE_PATH, 'sql', `${newFilename}.sql`), templates.sql);
     write(path.resolve(MIGRATIONS_BASE_PATH, `${newFilename}.js`), templates.js);
 
