@@ -2,7 +2,7 @@ module.exports = {
   coverageDirectory: '<rootDir>/coverage/integration',
   coveragePathIgnorePatterns: ['<rootDir>/build/', '<rootDir>/node_modules/', '<rootDir>/tests/utils/'],
   errorOnDeprecated: true,
-  globalSetup: '<rootDir>/tests/global_setup.js',
+  globalSetup: '<rootDir>/tests/global_setup.ts',
   testPathIgnorePatterns: [
     '<rootDir>/build/',
     '<rootDir>/tests/utils',
