@@ -24,13 +24,13 @@ export const visitActivitiesGetQuery = {
 };
 
 export const visitActivitiesPostPayload = {
-  name: Joi.string().required(),
-  category: Joi.string().required(),
+  name: Joi.string().trim().required(),
+  category: Joi.string().trim().required(),
 };
 
 export const visitActivitiesPutPayload = {
-  name: Joi.string().min(2),
-  category: Joi.string().min(3),
+  name: Joi.string().min(2).trim(),
+  category: Joi.string().min(3).trim(),
   monday: Joi.boolean(),
   tuesday: Joi.boolean(),
   wednesday: Joi.boolean(),
@@ -41,16 +41,26 @@ export const visitActivitiesPutPayload = {
 };
 
 export const cbPayload = {
-  name: Joi.string().min(1),
-  region: Joi.string().min(1),
-  sector: Joi.string().min(1),
-  logoUrl: Joi.string().uri(),
-  address1: Joi.string().min(1),
-  address2: Joi.string().min(1),
-  townCity: Joi.string().min(1),
-  postCode: Joi.string().min(6).max(10),
-  turnoverBand: Joi.string().min(5).max(11),
-  _360GivingId: Joi.string().min(5),
+  name: Joi.string().min(1).trim(),
+  region: Joi.string().min(1).trim(),
+  sector: Joi.string().min(1).trim(),
+  logoUrl: Joi.string().uri().trim(),
+  address1: Joi.string().min(1).trim(),
+  address2: Joi.string().min(1).trim(),
+  townCity: Joi.string().min(1).trim(),
+  postCode: Joi.string().min(6).max(10).trim(),
+  turnoverBand: Joi.string().min(5).max(11).trim(),
+  _360GivingId: Joi.string().min(5).trim(),
 };
 
 export const meOrId = id.allow('me');
+
+export const volunteerLogActivity = Joi.string().trim();
+
+export const volunteerLogDuration = Joi.object({
+  hours: Joi.number().integer().min(0),
+  minutes: Joi.number().integer().min(0),
+  seconds: Joi.number().integer().min(0),
+});
+
+export const volunteerProject = Joi.string().min(2).max(255);
