@@ -8,6 +8,7 @@ import { Dictionary, omit } from 'ramda';
 import { RegionEnum, SectorEnum } from '../types';
 import { CbAdmins } from '../cb_admin';
 import { RoleEnum, Roles } from '../../auth';
+import moment = require('moment');
 
 
 describe('Community Business Model', () => {
@@ -85,7 +86,7 @@ describe('Community Business Model', () => {
         { name: 'Black Mesa Research', frontlineWorkspaceId: null, frontlineApiKey: null }]));
     });
 
-    test('getOne :: returns first organisation only', async () => {
+    test('getOne :: returns first community business only', async () => {
       const org = await CommunityBusinesses.getOne(knex);
 
       expect(org).toEqual(expect.objectContaining({
@@ -94,21 +95,21 @@ describe('Community Business Model', () => {
       }));
     });
 
-    test('getOne :: returns null for non-existent organisation', async () => {
+    test('getOne :: returns null for non-existent community business', async () => {
       const org = await CommunityBusinesses.getOne(knex, { where: { id: 300 } });
       expect(org).toEqual(null);
     });
 
-    test('exists :: returns true for existent user', async () => {
+    test('exists :: returns true for existent community business', async () => {
       const exists = await CommunityBusinesses.exists(knex, {
         where: { name: 'Aperture Science' },
       });
       expect(exists).toBe(true);
     });
 
-    test('exists :: returns false for non-existent user', async () => {
+    test('exists :: returns false for non-existent community business', async () => {
       const exists = await CommunityBusinesses.exists(knex, {
-        where: { name: 'Umbrella Corporation' },
+        where: { name: 'Non-existent CB' },
       });
       expect(exists).toBe(false);
     });
@@ -315,6 +316,18 @@ describe('Community Business Model', () => {
 
       expect(logs.length).toEqual(0);
     });
+
+    test(':: returns subset of logs between dates', async () => {
+      const until = moment().toDate();
+      const since = moment().subtract(5, 'days').toDate();
+      const cb = await CommunityBusinesses.getOne(trx, { where: { id: 1 } });
+      const logs = await CommunityBusinesses
+        .getVisitLogsWithUsers(trx, cb, { since, until });
+
+      expect(logs).toHaveLength(6);
+      expect(logs.filter((l) => l.visitActivity === 'Free Running')).toHaveLength(4);
+      expect(logs.filter((l) => l.visitActivity === 'Wear Pink')).toHaveLength(2);
+    });
   });
 
   describe('getVisitActivities', () => {
@@ -363,6 +376,7 @@ describe('Community Business Model', () => {
       ]));
     });
   });
+
   describe('getAggregatedVisitLogs', () => {
     test(':: returns all aggregated logs for a cb', async () => {
       const cb = await CommunityBusinesses.getOne(trx, { where: { id: 1 } });
