@@ -306,6 +306,7 @@ type UsersBaseCollection = Collection<User> & {
 export type UserCollection = UsersBaseCollection & {
   createPasswordResetToken: (k: Knex, u: User) => Promise<SingleUseToken>
   usePasswordResetToken: (k: Knex, e: string, t: string) => Promise<null>
+  addActiveDayEvent: (k: Knex, u: User, o: string) => Promise<void>
 };
 
 export type VisitorCollection = UsersBaseCollection & {
@@ -353,8 +354,10 @@ export type CommunityBusinessCollection = Collection<CommunityBusiness> & {
   updateVisitActivity: (k: Knex, a: Partial<VisitActivity>) => Promise<Maybe<VisitActivity>>;
   deleteVisitActivity: (k: Knex, i: Int) => Promise<Maybe<VisitActivity>>;
   addVisitLog: (k: Knex, v: VisitActivity, u: Partial<User>) => Promise<VisitEvent>;
+  // TODO [getVisitLogsWithUsers]:
+  // this is still wrong, we return "category", which is missing
   getVisitLogsWithUsers: (k: Knex, c: CommunityBusiness, q?: ModelQuery<LinkedVisitEvent & User>) =>
-    Promise<Partial<LinkedVisitEvent & User>[]>;
+    Promise<Partial<LinkedVisitEvent & Pick<User, 'birthYear' | 'gender'>>[]>;
   getVisitLogAggregates: (
     k: Knex,
     c: CommunityBusiness,
